Merge remote-tracking branch 'remotes/chakracore/master' into WebAssembly

# Conflicts:
#	bin/ch/ChakraRtInterface.cpp
#	bin/ch/ChakraRtInterface.h
#	bin/ch/Helpers.cpp
#	bin/ch/WScriptJsrt.cpp
#	bin/ch/WScriptJsrt.h
#	lib/Jsrt/ChakraCommon.h
#	lib/Jsrt/Jsrt.cpp
#	lib/Runtime/Base/Chakra.Runtime.Base.vcxproj
#	lib/Runtime/Language/InterpreterLoop.inl
#	lib/Runtime/Library/InJavascript/Intl.js.bc.32b.h
#	lib/Runtime/Library/InJavascript/Intl.js.bc.64b.h
#	lib/Runtime/Library/InJavascript/Intl.js.nojit.bc.32b.h
#	lib/Runtime/Library/InJavascript/Intl.js.nojit.bc.64b.h
#	lib/Runtime/Runtime.h

+++ lib/Runtime/Library/JavascriptError.cpp
@@ -454,7 +454,7 @@ namespace Js
         pError->SetNotEnumerable(PropertyIds::description);
 
         hr = JavascriptError::GetErrorNumberFromResourceID(hr);
-        JavascriptOperators::InitProperty(pError, PropertyIds::number, JavascriptNumber::ToVar(hr, scriptContext));
+        JavascriptOperators::InitProperty(pError, PropertyIds::number, JavascriptNumber::ToVar((int32)hr, scriptContext));
         pError->SetNotEnumerable(PropertyIds::number);
     }
